Move get_document_by_entry() to DocumentStore
